+++ postgrest_py/__init__.py
@@ -1,6 +1,8 @@
 from __future__ import annotations
 
-__version__ = "0.7.1"
+__version__ = "0.8.1"
+
+from httpx import Timeout
 
 from ._async.client import AsyncPostgrestClient
 from ._async.request_builder import (