+++ postgrest_py/_sync/client.py
@@ -1,12 +1,16 @@
 from __future__ import annotations
 
-from typing import Dict, cast
+from typing import Dict, Union, cast
 
 from deprecation import deprecated
-from httpx import Response
+from httpx import Response, Timeout
 
 from .. import __version__
-from ..base_client import DEFAULT_POSTGREST_CLIENT_HEADERS, BasePostgrestClient
+from ..base_client import (
+    DEFAULT_POSTGREST_CLIENT_HEADERS,
+    DEFAULT_POSTGREST_CLIENT_TIMEOUT,
+    BasePostgrestClient,
+)
 from ..utils import SyncClient
 from .request_builder import SyncRequestBuilder
 
@@ -20,16 +24,28 @@ class SyncPostgrestClient(BasePostgrestClient):
         *,
         schema: str = "public",
         headers: Dict[str, str] = DEFAULT_POSTGREST_CLIENT_HEADERS,
+        timeout: Union[int, float, Timeout] = DEFAULT_POSTGREST_CLIENT_TIMEOUT,
     ) -> None:
-        BasePostgrestClient.__init__(self, base_url, schema=schema, headers=headers)
+        BasePostgrestClient.__init__(
+            self,
+            base_url,
+            schema=schema,
+            headers=headers,
+            timeout=timeout,
+        )
         self.session = cast(SyncClient, self.session)
 
     def create_session(
         self,
         base_url: str,
         headers: Dict[str, str],
+        timeout: Union[int, float, Timeout],
     ) -> SyncClient:
-        return SyncClient(base_url=base_url, headers=headers)
+        return SyncClient(
+            base_url=base_url,
+            headers=headers,
+            timeout=timeout,
+        )
 
     def __enter__(self) -> SyncPostgrestClient:
         return self
@@ -44,7 +60,7 @@ class SyncPostgrestClient(BasePostgrestClient):
         """Perform a table operation."""
         base_url = str(self.session.base_url)
         headers = dict(self.session.headers.items())
-        session = self.create_session(base_url, headers)
+        session = self.create_session(base_url, headers, self.session.timeout)
         session.auth = self.session.auth
         return SyncRequestBuilder(session, f"/{table}")